% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_wahlkreise.R
\name{add_wahlkreis}
\alias{add_wahlkreis}
\title{add_wahlkreis}
\usage{
add_wahlkreis(df)
}
\value{
the data frame with an additional column called Wahlkreis
}
\description{
A function that determines the "Wahlkreis" based on the "Stadtkreis" in the city of Zurich.
}
\examples{
data <- add_wahlkreis(df)
}
