% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_url.R
\name{build_quartierspiegel_url}
\alias{build_quartierspiegel_url}
\title{Build Quartierspiegel URL from a single-row data.frame}
\usage{
build_quartierspiegel_url(data)
}
\arguments{
\item{data}{A data.frame containing at least the columns:
quartier_nr}
}
\value{
A character string with the URL
}
\description{
Build Quartierspiegel URL from a single-row data.frame
}
