% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_create_url.R
\name{build_stadtplan_url}
\alias{build_stadtplan_url}
\title{Build Stadtplan URL from a single-row data.frame}
\usage{
build_stadtplan_url(data)
}
\arguments{
\item{data}{A data.frame containing at least the columns:
lokalisationsname, hausnummer, gebaeudeeingangnummer}
}
\value{
A character string with the URL
}
\description{
Build Stadtplan URL from a single-row data.frame
}
