% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sszadressengolem-package.R
\docType{package}
\name{sszadressengolem-package}
\alias{sszadressengolem}
\alias{sszadressengolem-package}
\title{sszadressengolem: Quartierfindeapp}
\description{
A Shiny app returns location information given an address in the city of Zurich.
}
\author{
\strong{Maintainer}: Sara Hänzi \email{sara.haenzi@zuerich.ch}

Authors:
\itemize{
  \item Sara Hänzi \email{sara.haenzi@zuerich.ch}
  \item Laura Näf \email{laura.naef@zuerich.ch}
  \item Alessandro Feller \email{alessandro.feller@zuerich.ch}
}

Other contributors:
\itemize{
  \item Statistik Stadt Zürich \email{statistik@zuerich.ch} [copyright holder, funder]
}

}
\keyword{internal}
